single-threaded compression done

+++ common/header.go
@@ -3,11 +3,11 @@ package common
 // The header for compressed files
 type HuffHeader struct {
 	magicNumber      uint32
-	treeSize         uint16
-	originalFileSize uint64
+	treeSize         uint32
+	originalFileSize int64
 }
 
 // CreateHeader creates a header
 func CreateHeader(treeSize uint16, originalFileSize uint64) *HuffHeader {
-	return &HuffHeader{MAGIC_NUMBER, treeSize, originalFileSize}
+	return &HuffHeader{MAGIC_NUMBER, uint32(treeSize), int64(originalFileSize)}
 }
